requirements.txt
tqdm==4.28.1
yacs==0.1.6
h5py==2.10.0
numpy==1.17.5
scipy==1.2.1
numba==0.47.0
smplx==0.1.13
gdown==3.6.4
PyYAML==5.3.1
joblib==0.14.1
pillow==6.2.1
trimesh==3.5.25
pyrender==0.1.36
progress==1.5
filterpy==1.4.5
matplotlib==3.3.2
tensorflow==1.15.4
tensorboard==1.15.0
torchvision==0.5.0
scikit-image==0.16.2
scikit-video==1.1.11
opencv-python==4.1.2.30
llvmlite==0.32.1
pymatreader==0.0.24
cdflib==0.3.19
neural_renderer_pytorch==1.1.3
git+https://github.com/mattloper/chumpy.git
git+https://github.com/mkocabas/yolov3-pytorch.git
git+https://github.com/mkocabas/multi-person-tracker.git
